chore(mobile): create src folder and move navigation/screens into structured layout
- Added src directory to apps/mobile for better code organization
- Prepared structure for navigation, screens, and UI components
- Kept domain logic separate in packages/domain for modularity
